agent/requirements.txt
websockets>=12.0
pyautogui>=0.9.54
Pillow>=10.0.0
